requirements.txt
streamlit
requests
python-dotenv
pydantic
Pillow
